moving out classes from base_util
moving out SPAWN_OBJECTS from base_util

Add mobs to memory

fix signature get_default_conditions

remove ununsed variables
