+++ agents/craftassist/tests/fake_agent.py
@@ -26,6 +26,7 @@ from droidlet.perception.craftassist.low_level_perception import LowLevelMCPerce
 from droidlet.perception.craftassist.heuristic_perception import PerceptionWrapper, check_inside
 from droidlet.perception.craftassist.rotation import look_vec, yaw_pitch
 from droidlet.interpreter.craftassist import dance
+from droidlet.lowlevel.minecraft.mc_util import SPAWN_OBJECTS
 
 # how many internal, non-world-interacting steps agent takes before world steps:
 WORLD_STEP = 10
@@ -312,6 +313,9 @@ class FakeAgent(LocoMCAgent):
     def init_memory(self):
         T = FakeMCTime(self.world)
         self.memory = MCAgentMemory(load_minecraft_specs=False, agent_time=T)
+        # Load mob types to memory
+        self.memory.load_mob_types(load_mob_types=True, spawn_objects=SPAWN_OBJECTS)
+        # Add dances to memory
         dance.add_default_dances(self.memory)
 
     def init_controller(self):
