+++ agents/loco_mc_agent.py
@@ -8,7 +8,7 @@ import time
 import numpy as np
 
 from agents.core import BaseAgent
-from droidlet.base_util import ErrorWithResponse
+from droidlet.shared_data_structs import ErrorWithResponse
 from droidlet.event import sio
 
 from droidlet.base_util import hash_user
