+++ droidlet/dialog/craftassist/dialogue_objects/mc_interpreter.py
@@ -41,7 +41,8 @@ from .spatial_reasoning import ComputeLocations
 from .condition_helper import MCConditionInterpreter
 from .attribute_helper import MCAttributeInterpreter
 from .point_target import PointTargetInterpreter
-from droidlet.base_util import ErrorWithResponse, number_from_span
+from droidlet.base_util import number_from_span
+from droidlet.shared_data_structs import ErrorWithResponse
 from droidlet.memory.memory_nodes import PlayerNode
 from droidlet.memory.craftassist.mc_memory_nodes import MobNode, ItemStackNode
 from droidlet.interpreter.craftassist import tasks, dance
