+++ droidlet/dialog/craftassist/dialogue_objects/spatial_reasoning.py
@@ -5,7 +5,8 @@ Copyright (c) Facebook, Inc. and its affiliates.
 import numpy as np
 from droidlet.perception.craftassist import rotation, shapes, heuristic_perception
 
-from droidlet.lowlevel.minecraft.mc_util import pos_to_np, to_block_center, to_block_pos, ErrorWithResponse
+from droidlet.lowlevel.minecraft.mc_util import pos_to_np, to_block_center, to_block_pos
+from droidlet.shared_data_structs import ErrorWithResponse
 
 DEFAULT_NUM_STEPS = 5
