+++ droidlet/dialog/robot/dialogue_objects/get_memory_handler.py
@@ -14,7 +14,7 @@ from droidlet.dialog.dialogue_objects import (
 )
 from .spatial_reasoning import ComputeLocations
 from .point_target import PointTargetInterpreter
-from droidlet.base_util import ErrorWithResponse
+from droidlet.shared_data_structs import ErrorWithResponse
 from droidlet.dialog.ttad.generation_dialogues.generate_utils import prepend_a_an
 from droidlet.interpreter.robot.tasks import Point
