+++ droidlet/lowlevel/minecraft/mc_util.py
@@ -213,3 +213,50 @@ def to_block_center(array):
 def to_block_pos(array):
     """Convert array to block position"""
     return np.floor(array).astype("int32")
+
+
+SPAWN_OBJECTS = {
+    "elder guardian": 4,
+    "wither skeleton": 5,
+    "stray": 6,
+    "husk": 23,
+    "zombie villager": 27,
+    "skeleton horse": 28,
+    "zombie horse": 29,
+    "donkey": 31,
+    "mule": 32,
+    "evoker": 34,
+    "vex": 35,
+    "vindicator": 36,
+    "creeper": 50,
+    "skeleton": 51,
+    "spider": 52,
+    "zombie": 54,
+    "slime": 55,
+    "ghast": 56,
+    "zombie pigman": 57,
+    "enderman": 58,
+    "cave spider": 59,
+    "silverfish": 60,
+    "blaze": 61,
+    "magma cube": 62,
+    "bat": 65,
+    "witch": 66,
+    "endermite": 67,
+    "guardian": 68,
+    "shulker": 69,
+    "pig": 90,
+    "sheep": 91,
+    "cow": 92,
+    "chicken": 93,
+    "squid": 94,
+    "wolf": 95,
+    "mooshroom": 96,
+    "ocelot": 98,
+    "horse": 100,
+    "rabbit": 101,
+    "polar bear": 102,
+    "llama": 103,
+    "parrot": 105,
+    "villager": 120,
+}
